split-utils * Move isNotNullOrUndefined to utils/typeguards.ts
