+++ src/dashboards/Compact/DepartureTile/DepartureTile.tsx
@@ -6,7 +6,6 @@ import {
     getTransportIconIdentifier,
     createTileSubLabel,
     getIconColorType,
-    isNotNullOrUndefined,
     filterMap,
 } from '../../../utils'
 import {
@@ -18,6 +17,7 @@ import { Tile } from '../components/Tile/Tile'
 import { TileRow } from '../components/TileRow/TileRow'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
+import { isNotNullOrUndefined } from '../../../utils/typeguards'
 
 function getTransportHeaderIcons(
     departures: LineData[],
