+++ src/logic/use-nearest-places/useNearestPlaces.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react'
 import { NearestPlace, Coordinates, TypeName } from '@entur/sdk'
-import { isNotNullOrUndefined } from '../../utils'
 import { apolloClient } from '../../apollo-client'
+import { isNotNullOrUndefined } from '../../utils/typeguards'
 import GET_NEAREST_PLACES_QUERY from './GetNearestPlaces.journey-planner.graphql'
 
 type GetNearestPlacesVariables = {
