+++ src/utils/typeguards.ts
@@ -1,0 +1,5 @@
+function isNotNullOrUndefined<T>(thing: T | undefined | null): thing is T {
+    return thing !== undefined && thing !== null
+}
+
+export { isNotNullOrUndefined }
